+++ ios/Plugin/Extensions/CapacitorVideoPlayerPlugin.Fullscreen.swift
@@ -17,16 +17,17 @@ extension CapacitorVideoPlayerPlugin {
     // swiftlint:disable function_parameter_count
     func createVideoPlayerFullscreenView(
         call: CAPPluginCall, videoUrl: URL, rate: Float,
-        exitOnEnd: Bool, loopOnEnd: Bool, subTitleUrl: URL?,
-        subTitleLanguage: String?, subTitleOptions: [String: Any]?) {
+        exitOnEnd: Bool, loopOnEnd: Bool, pipEnabled: Bool,
+        subTitleUrl: URL?, subTitleLanguage: String?,
+        subTitleOptions: [String: Any]?) {
         DispatchQueue.main.async { [weak self] in
             let playerId: String = self?.fsPlayerId ?? "fullscreen"
             if let fullscreenView = self?.implementation
                 .createFullscreenPlayer(
                     playerId: playerId, videoUrl: videoUrl,
                     rate: rate, exitOnEnd: exitOnEnd, loopOnEnd: loopOnEnd,
-                    subTitleUrl: subTitleUrl, language: subTitleLanguage,
-                    options: subTitleOptions) {
+                    pipEnabled: pipEnabled, subTitleUrl: subTitleUrl,
+                    language: subTitleLanguage, options: subTitleOptions) {
                 self?.videoPlayerFullScreenView = fullscreenView
                 self?.bgPlayer = self?.videoPlayerFullScreenView?.videoPlayer.player
                 guard let videoPlayer: AVPlayerViewController =